Refactor data blending interface
